Move JL parity tests into existing unit test files
Distribute Julia parity tests from tests/suites/jl_parity.rs into
the corresponding unit test files (15 model + 12 rule files). Tests
use shared JSON parsing helpers via include!("../jl_helpers.rs").
