+++ src/unit_tests/models/graph/maximal_is.rs
@@ -1,6 +1,7 @@
 use super::*;
 use crate::solvers::BruteForce;
 use crate::types::SolutionSize;
+include!("../../jl_helpers.rs");
 
 #[test]
 fn test_maximal_is_creation() {
@@ -223,4 +224,29 @@ fn test_maximal_is_problem() {
     // Not maximal: {0} alone - vertex 2 could be added
     assert_eq!(p.evaluate(&[1, 0, 0]), SolutionSize::Invalid);
     assert_eq!(p.direction(), Direction::Maximize);
+}
+
+#[test]
+fn test_jl_parity_evaluation() {
+    let data: serde_json::Value =
+        serde_json::from_str(include_str!("../../../../tests/data/jl/maximalis.json")).unwrap();
+    for instance in data["instances"].as_array().unwrap() {
+        let nv = instance["instance"]["num_vertices"].as_u64().unwrap() as usize;
+        let edges = jl_parse_edges(&instance["instance"]);
+        let problem = MaximalIS::<SimpleGraph, i32>::new(nv, edges);
+        for eval in instance["evaluations"].as_array().unwrap() {
+            let config = jl_parse_config(&eval["config"]);
+            let result = problem.evaluate(&config);
+            let jl_valid = eval["is_valid"].as_bool().unwrap();
+            assert_eq!(result.is_valid(), jl_valid, "MaximalIS validity mismatch for config {:?}", config);
+            if jl_valid {
+                let jl_size = eval["size"].as_i64().unwrap() as i32;
+                assert_eq!(result.unwrap(), jl_size, "MaximalIS size mismatch for config {:?}", config);
+            }
+        }
+        let best = BruteForce::new().find_all_best(&problem);
+        let jl_best = jl_parse_configs_set(&instance["best_solutions"]);
+        let rust_best: HashSet<Vec<usize>> = best.into_iter().collect();
+        assert_eq!(rust_best, jl_best, "MaximalIS best solutions mismatch");
+    }
 }
