+++ src/unit_tests/models/graph/maximum_independent_set.rs
@@ -2,6 +2,7 @@ use super::*;
 use crate::solvers::BruteForce;
 use crate::traits::{OptimizationProblem, Problem};
 use crate::types::{Direction, SolutionSize};
+include!("../../jl_helpers.rs");
 
 #[test]
 fn test_independent_set_creation() {
@@ -248,4 +249,34 @@ fn test_problem_name() {
         <MaximumIndependentSet<SimpleGraph, i32> as Problem>::NAME,
         "MaximumIndependentSet"
     );
+}
+
+#[test]
+fn test_jl_parity_evaluation() {
+    let data: serde_json::Value =
+        serde_json::from_str(include_str!("../../../../tests/data/jl/independentset.json")).unwrap();
+    for instance in data["instances"].as_array().unwrap() {
+        let nv = instance["instance"]["num_vertices"].as_u64().unwrap() as usize;
+        let edges = jl_parse_edges(&instance["instance"]);
+        let weights = jl_parse_i32_vec(&instance["instance"]["weights"]);
+        let problem = if weights.iter().all(|&w| w == 1) {
+            MaximumIndependentSet::<SimpleGraph, i32>::new(nv, edges)
+        } else {
+            MaximumIndependentSet::with_weights(nv, edges, weights)
+        };
+        for eval in instance["evaluations"].as_array().unwrap() {
+            let config = jl_parse_config(&eval["config"]);
+            let result = problem.evaluate(&config);
+            let jl_valid = eval["is_valid"].as_bool().unwrap();
+            assert_eq!(result.is_valid(), jl_valid, "IS validity mismatch for config {:?}", config);
+            if jl_valid {
+                let jl_size = eval["size"].as_i64().unwrap() as i32;
+                assert_eq!(result.unwrap(), jl_size, "IS size mismatch for config {:?}", config);
+            }
+        }
+        let best = BruteForce::new().find_all_best(&problem);
+        let jl_best = jl_parse_configs_set(&instance["best_solutions"]);
+        let rust_best: HashSet<Vec<usize>> = best.into_iter().collect();
+        assert_eq!(rust_best, jl_best, "IS best solutions mismatch");
+    }
 }
