+++ src/unit_tests/models/set/maximum_set_packing.rs
@@ -2,6 +2,7 @@ use super::*;
 use crate::solvers::BruteForce;
 use crate::traits::{OptimizationProblem, Problem};
 use crate::types::{Direction, SolutionSize};
+include!("../../jl_helpers.rs");
 
 #[test]
 fn test_set_packing_creation() {
@@ -192,4 +193,33 @@ fn test_set_packing_problem() {
     assert_eq!(Problem::evaluate(&p, &[0, 0, 0]), SolutionSize::Valid(0));
 
     assert_eq!(p.direction(), Direction::Maximize);
+}
+
+#[test]
+fn test_jl_parity_evaluation() {
+    let data: serde_json::Value =
+        serde_json::from_str(include_str!("../../../../tests/data/jl/setpacking.json")).unwrap();
+    for instance in data["instances"].as_array().unwrap() {
+        let sets = jl_parse_sets(&instance["instance"]["sets"]);
+        let weights = jl_parse_i32_vec(&instance["instance"]["weights"]);
+        let problem = if weights.iter().all(|&w| w == 1) {
+            MaximumSetPacking::<i32>::new(sets)
+        } else {
+            MaximumSetPacking::with_weights(sets, weights)
+        };
+        for eval in instance["evaluations"].as_array().unwrap() {
+            let config = jl_parse_config(&eval["config"]);
+            let result = problem.evaluate(&config);
+            let jl_valid = eval["is_valid"].as_bool().unwrap();
+            assert_eq!(result.is_valid(), jl_valid, "SetPacking validity mismatch for config {:?}", config);
+            if jl_valid {
+                let jl_size = eval["size"].as_i64().unwrap() as i32;
+                assert_eq!(result.unwrap(), jl_size, "SetPacking size mismatch for config {:?}", config);
+            }
+        }
+        let best = BruteForce::new().find_all_best(&problem);
+        let jl_best = jl_parse_configs_set(&instance["best_solutions"]);
+        let rust_best: HashSet<Vec<usize>> = best.into_iter().collect();
+        assert_eq!(rust_best, jl_best, "SetPacking best solutions mismatch");
+    }
 }
