+++ src/unit_tests/models/set/minimum_set_covering.rs
@@ -2,6 +2,7 @@ use super::*;
 use crate::solvers::BruteForce;
 use crate::traits::{OptimizationProblem, Problem};
 use crate::types::{Direction, SolutionSize};
+include!("../../jl_helpers.rs");
 
 #[test]
 fn test_set_covering_creation() {
@@ -173,4 +174,30 @@ fn test_set_covering_problem() {
     assert_eq!(Problem::evaluate(&p, &[0, 0]), SolutionSize::Invalid);
 
     assert_eq!(p.direction(), Direction::Minimize);
+}
+
+#[test]
+fn test_jl_parity_evaluation() {
+    let data: serde_json::Value =
+        serde_json::from_str(include_str!("../../../../tests/data/jl/setcovering.json")).unwrap();
+    for instance in data["instances"].as_array().unwrap() {
+        let universe_size = instance["instance"]["universe_size"].as_u64().unwrap() as usize;
+        let sets = jl_parse_sets(&instance["instance"]["sets"]);
+        let weights = jl_parse_i32_vec(&instance["instance"]["weights"]);
+        let problem = MinimumSetCovering::<i32>::with_weights(universe_size, sets, weights);
+        for eval in instance["evaluations"].as_array().unwrap() {
+            let config = jl_parse_config(&eval["config"]);
+            let result = problem.evaluate(&config);
+            let jl_valid = eval["is_valid"].as_bool().unwrap();
+            assert_eq!(result.is_valid(), jl_valid, "SetCovering validity mismatch for config {:?}", config);
+            if jl_valid {
+                let jl_size = eval["size"].as_i64().unwrap() as i32;
+                assert_eq!(result.unwrap(), jl_size, "SetCovering size mismatch for config {:?}", config);
+            }
+        }
+        let best = BruteForce::new().find_all_best(&problem);
+        let jl_best = jl_parse_configs_set(&instance["best_solutions"]);
+        let rust_best: HashSet<Vec<usize>> = best.into_iter().collect();
+        assert_eq!(rust_best, jl_best, "SetCovering best solutions mismatch");
+    }
 }
