+++ src/unit_tests/rules/sat_coloring.rs
@@ -1,6 +1,7 @@
 use super::*;
 use crate::models::satisfiability::CNFClause;
 use crate::solvers::BruteForce;
+include!("../jl_helpers.rs");
 
 #[test]
 fn test_constructor_basic_structure() {
@@ -292,4 +293,32 @@ fn test_extraction_with_different_color_assignment() {
     let coloring_permuted2 = vec![1, 2, 0, 1, 2];
     let extracted2 = reduction.extract_solution(&coloring_permuted2);
     assert_eq!(extracted2, vec![1]);
+}
+
+#[test]
+fn test_jl_parity_sat_to_coloring() {
+    let sat_data: serde_json::Value =
+        serde_json::from_str(include_str!("../../../tests/data/jl/satisfiability.json")).unwrap();
+    let fixtures: &[(&str, &str)] = &[
+        (include_str!("../../../tests/data/jl/satisfiability_to_coloring3.json"), "simple_clause"),
+        (include_str!("../../../tests/data/jl/rule_satisfiability2_to_coloring3.json"), "rule_sat_coloring"),
+    ];
+    for (fixture_str, label) in fixtures {
+        let data: serde_json::Value = serde_json::from_str(fixture_str).unwrap();
+        let inst = &jl_find_instance_by_label(&sat_data, label)["instance"];
+        let (num_vars, clauses) = jl_parse_sat_clauses(inst);
+        let source = Satisfiability::new(num_vars, clauses);
+        let result = ReduceTo::<KColoring<3, SimpleGraph>>::reduce_to(&source);
+        let ilp_solver = crate::solvers::ILPSolver::new();
+        let target = result.target_problem();
+        let target_sol = ilp_solver.solve_reduced(target).expect("ILP should find a coloring");
+        let extracted = result.extract_solution(&target_sol);
+        let best_source: HashSet<Vec<usize>> = BruteForce::new()
+            .find_all_satisfying(&source).into_iter().collect();
+        assert!(best_source.contains(&extracted), "SAT->Coloring [{label}]: extracted not satisfying");
+        for case in data["cases"].as_array().unwrap() {
+            assert_eq!(best_source, jl_parse_configs_set(&case["best_source"]),
+                "SAT->Coloring [{label}]: best source mismatch");
+        }
+    }
 }
