+++ src/unit_tests/rules/spinglass_maxcut.rs
@@ -1,5 +1,6 @@
 use super::*;
 use crate::solvers::BruteForce;
+include!("../jl_helpers.rs");
 
 #[test]
 fn test_maxcut_to_spinglass() {
@@ -89,4 +90,93 @@ fn test_reduction_structure() {
     let mc2 = reduction2.target_problem();
 
     assert_eq!(mc2.num_vertices(), 3);
+}
+
+#[test]
+fn test_jl_parity_spinglass_to_maxcut() {
+    let data: serde_json::Value =
+        serde_json::from_str(include_str!("../../../tests/data/jl/spinglass_to_maxcut.json")).unwrap();
+    let sg_data: serde_json::Value =
+        serde_json::from_str(include_str!("../../../tests/data/jl/spinglass.json")).unwrap();
+    let inst = &sg_data["instances"][0]["instance"];
+    let nv = inst["num_vertices"].as_u64().unwrap() as usize;
+    let edges = jl_parse_edges(inst);
+    let j_values = jl_parse_i32_vec(&inst["J"]);
+    let h_values = jl_parse_i32_vec(&inst["h"]);
+    let interactions: Vec<((usize, usize), i32)> = edges.into_iter().zip(j_values).collect();
+    let source = SpinGlass::<SimpleGraph, i32>::new(nv, interactions, h_values);
+    let result = ReduceTo::<MaxCut<SimpleGraph, i32>>::reduce_to(&source);
+    let solver = BruteForce::new();
+    let best_target = solver.find_all_best(result.target_problem());
+    let best_source: HashSet<Vec<usize>> = solver.find_all_best(&source).into_iter().collect();
+    let extracted: HashSet<Vec<usize>> = best_target.iter().map(|t| result.extract_solution(t)).collect();
+    assert!(extracted.is_subset(&best_source));
+    for case in data["cases"].as_array().unwrap() {
+        assert_eq!(best_source, jl_parse_configs_set(&case["best_source"]));
+    }
+}
+
+#[test]
+fn test_jl_parity_maxcut_to_spinglass() {
+    let data: serde_json::Value =
+        serde_json::from_str(include_str!("../../../tests/data/jl/maxcut_to_spinglass.json")).unwrap();
+    let mc_data: serde_json::Value =
+        serde_json::from_str(include_str!("../../../tests/data/jl/maxcut.json")).unwrap();
+    let inst = &mc_data["instances"][0]["instance"];
+    let nv = inst["num_vertices"].as_u64().unwrap() as usize;
+    let weighted_edges = jl_parse_weighted_edges(inst);
+    let source = MaxCut::<SimpleGraph, i32>::new(nv, weighted_edges);
+    let result = ReduceTo::<SpinGlass<SimpleGraph, i32>>::reduce_to(&source);
+    let solver = BruteForce::new();
+    let best_target = solver.find_all_best(result.target_problem());
+    let best_source: HashSet<Vec<usize>> = solver.find_all_best(&source).into_iter().collect();
+    let extracted: HashSet<Vec<usize>> = best_target.iter().map(|t| result.extract_solution(t)).collect();
+    assert!(extracted.is_subset(&best_source));
+    for case in data["cases"].as_array().unwrap() {
+        assert_eq!(best_source, jl_parse_configs_set(&case["best_source"]));
+    }
+}
+
+#[test]
+fn test_jl_parity_rule_maxcut_to_spinglass() {
+    let data: serde_json::Value =
+        serde_json::from_str(include_str!("../../../tests/data/jl/rule_maxcut_to_spinglass.json")).unwrap();
+    let mc_data: serde_json::Value =
+        serde_json::from_str(include_str!("../../../tests/data/jl/maxcut.json")).unwrap();
+    let inst = &jl_find_instance_by_label(&mc_data, "rule_4vertex")["instance"];
+    let source = MaxCut::<SimpleGraph, i32>::new(
+        inst["num_vertices"].as_u64().unwrap() as usize, jl_parse_weighted_edges(inst));
+    let result = ReduceTo::<SpinGlass<SimpleGraph, i32>>::reduce_to(&source);
+    let solver = BruteForce::new();
+    let best_target = solver.find_all_best(result.target_problem());
+    let best_source: HashSet<Vec<usize>> = solver.find_all_best(&source).into_iter().collect();
+    let extracted: HashSet<Vec<usize>> = best_target.iter().map(|t| result.extract_solution(t)).collect();
+    assert!(extracted.is_subset(&best_source));
+    for case in data["cases"].as_array().unwrap() {
+        assert_eq!(best_source, jl_parse_configs_set(&case["best_source"]));
+    }
+}
+
+#[test]
+fn test_jl_parity_rule_spinglass_to_maxcut() {
+    let data: serde_json::Value =
+        serde_json::from_str(include_str!("../../../tests/data/jl/rule_spinglass_to_maxcut.json")).unwrap();
+    let sg_data: serde_json::Value =
+        serde_json::from_str(include_str!("../../../tests/data/jl/spinglass.json")).unwrap();
+    let inst = &jl_find_instance_by_label(&sg_data, "rule_4vertex")["instance"];
+    let nv = inst["num_vertices"].as_u64().unwrap() as usize;
+    let edges = jl_parse_edges(inst);
+    let j_values = jl_parse_i32_vec(&inst["J"]);
+    let h_values = jl_parse_i32_vec(&inst["h"]);
+    let interactions: Vec<((usize, usize), i32)> = edges.into_iter().zip(j_values).collect();
+    let source = SpinGlass::<SimpleGraph, i32>::new(nv, interactions, h_values);
+    let result = ReduceTo::<MaxCut<SimpleGraph, i32>>::reduce_to(&source);
+    let solver = BruteForce::new();
+    let best_target = solver.find_all_best(result.target_problem());
+    let best_source: HashSet<Vec<usize>> = solver.find_all_best(&source).into_iter().collect();
+    let extracted: HashSet<Vec<usize>> = best_target.iter().map(|t| result.extract_solution(t)).collect();
+    assert!(extracted.is_subset(&best_source));
+    for case in data["cases"].as_array().unwrap() {
+        assert_eq!(best_source, jl_parse_configs_set(&case["best_source"]));
+    }
 }
